@@ -1,49 +1,64 @@
 {
-  "name": "yajra/laravel-datatables-html",
-  "description": "Laravel DataTables HTML builder plugin for Laravel 5.4+.",
-  "keywords": [
-    "laravel",
-    "dataTables",
-    "jquery",
-    "html",
-    "js"
-  ],
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Arjay Angeles",
-      "email": "aqangeles@gmail.com"
-    }
-  ],
-  "require": {
-    "php": "^8.1",
-    "ext-json": "*",
-    "yajra/laravel-datatables-oracle": "^10.0"
-  },
-  "require-dev": {
-    "nunomaduro/larastan": "^2.4",
-    "orchestra/testbench": "^7.21"
-  },
-  "autoload": {
-    "psr-4": {
-      "Yajra\\DataTables\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Yajra\\DataTables\\Html\\Tests\\": "tests/"
-    }
-  },
-  "extra": {
-    "branch-alias": {
-      "dev-master": "10.0-dev"
+    "name": "yajra/laravel-datatables-html",
+    "description": "Laravel DataTables HTML builder plugin",
+    "keywords": [
+        "yajra",
+        "laravel",
+        "dataTables",
+        "jquery",
+        "html",
+        "js"
+    ],
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Arjay Angeles",
+            "email": "aqangeles@gmail.com"
+        }
+    ],
+    "require": {
+        "php": "^8.2",
+        "ext-json": "*",
+        "yajra/laravel-datatables-oracle": "^11.0"
     },
-    "laravel": {
-      "providers": [
-        "Yajra\\DataTables\\HtmlServiceProvider"
-      ]
-    }
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true
+    "require-dev": {
+        "larastan/larastan": "^2.9.1",
+        "orchestra/testbench": "^9",
+        "laravel/pint": "^1.14",
+        "rector/rector": "^1.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Yajra\\DataTables\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Yajra\\DataTables\\Html\\Tests\\": "tests/"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "11.0-dev"
+        },
+        "laravel": {
+            "providers": [
+                "Yajra\\DataTables\\HtmlServiceProvider"
+            ]
+        }
+    },
+    "scripts": {
+        "test": "./vendor/bin/phpunit",
+        "pint": "./vendor/bin/pint",
+        "rector": "./vendor/bin/rector",
+        "stan": "./vendor/bin/phpstan analyse --memory-limit=2G --ansi --no-progress --no-interaction --configuration=phpstan.neon.dist",
+        "pr": [
+            "@pint",
+            "@rector",
+            "@stan",
+            "@test"
+        ]
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
