@@ -3,6 +3,7 @@
 namespace Yajra\DataTables\Html\Tests;
 
 use Illuminate\Support\HtmlString;
+use PHPUnit\Framework\Attributes\Test;
 use Yajra\DataTables\Html\Builder;
 use Yajra\DataTables\Html\Column;
 use Yajra\DataTables\Html\ColumnDefinition;
@@ -11,23 +12,23 @@ use Yajra\DataTables\Html\Editor\Editor;
 
 class BuilderTest extends TestCase
 {
-     /** @test */
-     public function it_can_get_script_default_type_attribute()
-     {
-         $html = $this->getHtmlBuilder()->scripts()->toHtml();
+    #[Test]
+    public function it_can_get_script_default_type_attribute()
+    {
+        $html = $this->getHtmlBuilder()->scripts()->toHtml();
 
-         $this->assertStringContainsString('type="text/javascript"', $html);
-     }
+        $this->assertStringContainsString('type="text/javascript"', $html);
+    }
 
-    /** @test */
-     public function it_can_set_script_type_attribute()
-     {
+    #[Test]
+    public function it_can_set_script_type_attribute()
+    {
         $html = $this->getHtmlBuilder()->scripts(attributes: ['type' => 'module'])->toHtml();
 
         $this->assertStringContainsString('type="module"', $html);
-     }
+    }
 
-     /** @test */
+    #[Test]
     public function it_can_set_multiple_script_attributes()
     {
         $html = $this->getHtmlBuilder()->scripts(attributes: ['prop1' => 'val1', 'prop2' => 'val2'])->toHtml();
@@ -36,7 +37,7 @@ class BuilderTest extends TestCase
         $this->assertStringContainsString('prop2="val2"', $html);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_use_vitejs_module_script()
     {
         Builder::useVite();
@@ -46,7 +47,7 @@ class BuilderTest extends TestCase
         Builder::useWebpack();
     }
 
-    /** @test */
+    #[Test]
     public function it_can_resolved_builder_class()
     {
         $builder = $this->getHtmlBuilder();
@@ -57,7 +58,7 @@ class BuilderTest extends TestCase
         $this->assertInstanceOf(Builder::class, $builder);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_read_table_id_from_config()
     {
         $this->assertEquals('dataTableBuilder', $this->getHtmlBuilder()->getTableId());
@@ -67,7 +68,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('test', $this->getHtmlBuilder()->getTableId());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_change_namespace()
     {
         $builder = $this->getHtmlBuilder();
@@ -79,16 +80,16 @@ class BuilderTest extends TestCase
         $this->assertStringContainsString('TestDataTables', $builder->scripts()->toHtml());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_generate_table_html_and_scripts()
     {
         $builder = $this->getHtmlBuilder();
 
         $builder->setTableId('foo-table')
-                ->columns([
-                    Column::make('foo'),
-                    Column::make('baz'),
-                ]);
+            ->columns([
+                Column::make('foo'),
+                Column::make('baz'),
+            ]);
 
         $table = $builder->table()->toHtml();
 
@@ -103,7 +104,7 @@ class BuilderTest extends TestCase
         $this->assertEquals($expected, $builder->generateScripts()->toHtml());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_set_table_attribute()
     {
         $builder = $this->getHtmlBuilder();
@@ -113,7 +114,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('val', $builder->getTableAttribute('attr'));
     }
 
-    /** @test */
+    #[Test]
     public function it_can_set_table_id_attribute()
     {
         $builder = $this->getHtmlBuilder();
@@ -123,7 +124,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('val', $builder->getTableAttribute('id'));
     }
 
-    /** @test */
+    #[Test]
     public function it_can_set_multiple_table_attributes()
     {
         $builder = $this->getHtmlBuilder();
@@ -134,7 +135,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('val2', $builder->getTableAttribute('prop2'));
     }
 
-    /** @test */
+    #[Test]
     public function it_can_get_inexistent_table_attribute_throws()
     {
         $builder = $this->getHtmlBuilder();
@@ -144,7 +145,7 @@ class BuilderTest extends TestCase
         $this->assertEmpty($attr);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_add_table_class_attribute()
     {
         $builder = $this->getHtmlBuilder();
@@ -160,7 +161,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('table foo bar a-b c value', $builder->getTableAttribute('class'));
     }
 
-    /** @test */
+    #[Test]
     public function it_can_remove_table_class_attribute()
     {
         $builder = $this->getHtmlBuilder();
@@ -176,7 +177,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('a', $builder->getTableAttribute('class'));
     }
 
-    /** @test */
+    #[Test]
     public function it_can_add_checkbox()
     {
         $builder = $this->getHtmlBuilder();
@@ -192,7 +193,7 @@ class BuilderTest extends TestCase
         $this->assertEquals(true, $column->printable);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_add_index_column()
     {
         $builder = $this->getHtmlBuilder();
@@ -208,7 +209,7 @@ class BuilderTest extends TestCase
         $this->assertEquals(true, $column->printable);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_add_action_column()
     {
         $builder = $this->getHtmlBuilder();
@@ -224,7 +225,7 @@ class BuilderTest extends TestCase
         $this->assertEquals(true, $column->printable);
     }
 
-    /** @test */
+    #[Test]
     public function it_has_column_defs()
     {
         $builder = $this->getHtmlBuilder();
@@ -247,7 +248,7 @@ class BuilderTest extends TestCase
         $this->assertCount(1, $builder->getAttribute('columnDefs'));
     }
 
-    /** @test */
+    #[Test]
     public function it_has_table_options()
     {
         $builder = $this->getHtmlBuilder();
@@ -272,7 +273,7 @@ class BuilderTest extends TestCase
         $this->assertEquals('<table class="table" id="my-table"><thead class="thead-dark"><tr></tr></thead></table>', $builder->table()->toHtml());
     }
 
-    /** @test */
+    #[Test]
     public function it_has_editors()
     {
         $builder = $this->getHtmlBuilder();
@@ -286,6 +287,4 @@ class BuilderTest extends TestCase
         ]);
         $this->assertCount(2, $builder->getEditors());
     }
-
-
 }
