@@ -4,6 +4,7 @@ namespace Yajra\DataTables\Html\Tests;
 
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Illuminate\Support\Facades\DB;
+use PHPUnit\Framework\Attributes\Test;
 use Yajra\DataTables\Html\Editor\Fields\Options;
 use Yajra\DataTables\Html\Tests\Models\User;
 
@@ -11,7 +12,7 @@ class FieldOptionsTest extends TestCase
 {
     use DatabaseTransactions;
 
-    /** @test */
+    #[Test]
     public function it_has_true_false()
     {
         $options = Options::trueFalse();
@@ -22,7 +23,7 @@ class FieldOptionsTest extends TestCase
         ], $options->toArray());
     }
 
-    /** @test */
+    #[Test]
     public function it_has_yes_no()
     {
         $options = Options::yesNo();
@@ -33,7 +34,7 @@ class FieldOptionsTest extends TestCase
         ], $options->toArray());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_append_and_prepend()
     {
         $options = Options::yesNo();
@@ -59,28 +60,28 @@ class FieldOptionsTest extends TestCase
         ], $options->toArray());
     }
 
-    /** @test */
+    #[Test]
     public function it_can_get_options_from_table()
     {
         $options = Options::table('users', 'name');
         $this->assertCount(20, $options);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_get_options_from_query()
     {
         $options = Options::table(DB::table('users')->where('id', 1), 'name');
         $this->assertCount(1, $options);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_get_options_from_model()
     {
         $options = Options::model(User::class, 'name');
         $this->assertCount(20, $options);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_get_options_from_model_builder()
     {
         $options = Options::model(User::query()->whereKey(1), 'name');
