@@ -29,12 +29,13 @@ use Yajra\DataTables\Html\Options\Plugins\SearchPanes;
  * @property string $contentPadding
  * @property string $createdCell
  * @property string $exportFormat
+ *
  * @see https://datatables.net/reference/option/#columns
  */
 class Column extends Fluent
 {
-    use SearchPanes;
     use HasAuthorizations;
+    use SearchPanes;
 
     /**
      * @param  array  $attributes
@@ -66,9 +67,6 @@ class Column extends Fluent
 
     /**
      * Format string to title case.
-     *
-     * @param  string  $value
-     * @return string
      */
     public static function titleFormat(string $value): string
     {
@@ -78,8 +76,8 @@ class Column extends Fluent
     /**
      * Set column title.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.title
      */
     public function title(string $value): static
@@ -91,12 +89,8 @@ class Column extends Fluent
 
     /**
      * Create a computed column that is not searchable/orderable.
-     *
-     * @param  string  $data
-     * @param  string|null  $title
-     * @return Column
      */
-    public static function computed(string $data, string $title = null): Column
+    public static function computed(string $data, ?string $title = null): Column
     {
         if (is_null($title)) {
             $title = self::titleFormat($data);
@@ -108,8 +102,8 @@ class Column extends Fluent
     /**
      * Set column searchable flag.
      *
-     * @param  bool  $flag
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.searchable
      */
     public function searchable(bool $flag = true): static
@@ -122,8 +116,8 @@ class Column extends Fluent
     /**
      * Set column orderable flag.
      *
-     * @param  bool  $flag
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.orderable
      */
     public function orderable(bool $flag = true): static
@@ -135,10 +129,6 @@ class Column extends Fluent
 
     /**
      * Make a new column instance.
-     *
-     * @param  array|string  $data
-     * @param  string  $name
-     * @return static
      */
     public static function make(array|string $data = [], string $name = ''): static
     {
@@ -155,9 +145,6 @@ class Column extends Fluent
 
     /**
      * Make a new formatted column instance.
-     *
-     * @param  string  $name
-     * @return static
      */
     public static function formatted(string $name): static
     {
@@ -173,25 +160,21 @@ class Column extends Fluent
 
     /**
      * Create a checkbox column.
-     *
-     * @param  string  $title
-     * @return static
      */
     public static function checkbox(string $title = ''): static
     {
         return static::make('')
-                     ->content('')
-                     ->title($title)
-                     ->className('select-checkbox')
-                     ->orderable(false)
-                     ->exportable(false)
-                     ->searchable(false);
+            ->content('')
+            ->title($title)
+            ->className('select-checkbox')
+            ->orderable(false)
+            ->exportable(false)
+            ->searchable(false);
     }
 
     /**
      * Set column exportable flag.
      *
-     * @param  bool  $flag
      * @return $this
      */
     public function exportable(bool $flag = true): static
@@ -204,8 +187,8 @@ class Column extends Fluent
     /**
      * Set column class name.
      *
-     * @param  string  $class
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.className
      */
     public function className(string $class): static
@@ -218,8 +201,8 @@ class Column extends Fluent
     /**
      * Set column default content.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.defaultContent
      */
     public function content(string $value): static
@@ -232,8 +215,8 @@ class Column extends Fluent
     /**
      * Set column responsive priority.
      *
-     * @param  int|string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.responsivePriority
      */
     public function responsivePriority(int|string $value): static
@@ -247,6 +230,7 @@ class Column extends Fluent
      * Set column hidden state.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.visible
      */
     public function hidden(): static
@@ -257,8 +241,8 @@ class Column extends Fluent
     /**
      * Set column visible flag.
      *
-     * @param  bool  $flag
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.visible
      */
     public function visible(bool $flag = true): static
@@ -271,7 +255,6 @@ class Column extends Fluent
     /**
      * Append a class name to field.
      *
-     * @param  string  $class
      * @return $this
      */
     public function addClass(string $class): static
@@ -288,7 +271,6 @@ class Column extends Fluent
     /**
      * Set column printable flag.
      *
-     * @param  bool  $flag
      * @return $this
      */
     public function printable(bool $flag = true): static
@@ -301,8 +283,8 @@ class Column extends Fluent
     /**
      * Set column width value.
      *
-     * @param  int|string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.width
      */
     public function width(int|string $value): static
@@ -315,8 +297,8 @@ class Column extends Fluent
     /**
      * Set column data option value.
      *
-     * @param  array|string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.data
      * @see https://datatables.net/manual/data/orthogonal-data
      */
@@ -330,8 +312,8 @@ class Column extends Fluent
     /**
      * Set column name option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.name
      */
     public function name(string $value): static
@@ -344,8 +326,8 @@ class Column extends Fluent
     /**
      * Set column edit field option value.
      *
-     * @param  array|string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.editField
      */
     public function editField(array|string $value): static
@@ -358,8 +340,8 @@ class Column extends Fluent
     /**
      * Set column orderData option value.
      *
-     * @param  array|int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.orderData
      */
     public function orderData(array|int $value): static
@@ -372,8 +354,8 @@ class Column extends Fluent
     /**
      * Set column orderDataType option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.orderDataType
      */
     public function orderDataType(string $value): static
@@ -386,8 +368,8 @@ class Column extends Fluent
     /**
      * Set column orderSequence option value.
      *
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.orderSequence
      */
     public function orderSequence(array $value): static
@@ -400,8 +382,8 @@ class Column extends Fluent
     /**
      * Set column cellType option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.cellType
      */
     public function cellType(string $value = 'th'): static
@@ -414,8 +396,8 @@ class Column extends Fluent
     /**
      * Set column type option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.type
      */
     public function type(string $value): static
@@ -428,8 +410,8 @@ class Column extends Fluent
     /**
      * Set column contentPadding option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.contentPadding
      */
     public function contentPadding(string $value): static
@@ -442,8 +424,8 @@ class Column extends Fluent
     /**
      * Set column createdCell option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.createdCell
      */
     public function createdCell(string $value): static
@@ -456,9 +438,9 @@ class Column extends Fluent
     /**
      * Use the js renderer "$.fn.dataTable.render.".
      *
-     * @param  string  $value
      * @param  int|string|null  ...$params
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.render
      */
     public function renderJs(string $value, ...$params): static
@@ -480,8 +462,8 @@ class Column extends Fluent
     /**
      * Set column renderer.
      *
-     * @param  mixed  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.render
      */
     public function render(mixed $value): static
@@ -494,9 +476,7 @@ class Column extends Fluent
     /**
      * Set Callback function to render column for Print + Export
      *
-     * @param  callable  $callback
      * @return $this
-
      */
     public function exportRender(callable $callback): static
     {
@@ -507,9 +487,6 @@ class Column extends Fluent
 
     /**
      * Parse render attribute.
-     *
-     * @param  mixed  $value
-     * @return string|null
      */
     public function parseRender(mixed $value): ?string
     {
@@ -526,7 +503,7 @@ class Column extends Fluent
             return $value($parameters);
         } elseif ($this->isBuiltInRenderFunction($value)) {
             return $value;
-        } elseif (strlen($value) < 256 && $view->exists($value)) {
+        } elseif (strlen((string) $value) < 256 && $view->exists($value)) {
             return $view->make($value)->with($parameters)->render();
         }
 
@@ -535,9 +512,6 @@ class Column extends Fluent
 
     /**
      * Check if given key & value is a valid datatables built-in renderer function.
-     *
-     * @param  string  $value
-     * @return bool
      */
     private function isBuiltInRenderFunction(string $value): bool
     {
@@ -550,9 +524,6 @@ class Column extends Fluent
 
     /**
      * Display render value as is.
-     *
-     * @param  string  $value
-     * @return string
      */
     private function parseRenderAsString(string $value): string
     {
@@ -562,8 +533,8 @@ class Column extends Fluent
     /**
      * Set column renderer with give raw value.
      *
-     * @param  mixed  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.render
      */
     public function renderRaw(mixed $value): static
@@ -576,7 +547,6 @@ class Column extends Fluent
     /**
      * Set column footer.
      *
-     * @param  mixed  $value
      * @return $this
      */
     public function footer(mixed $value): static
@@ -589,7 +559,6 @@ class Column extends Fluent
     /**
      * Set custom html title instead default label.
      *
-     * @param  mixed  $value
      * @return $this
      */
     public function titleAttr(mixed $value): static
@@ -602,8 +571,8 @@ class Column extends Fluent
     /**
      * Set excel column format when exporting.
      *
-     * @param  string|callable $format
      * @return $this
+     *
      * @see https://github.com/yajra/laravel-datatables-export
      */
     public function exportFormat(string|callable $format): static
@@ -613,9 +582,6 @@ class Column extends Fluent
         return $this;
     }
 
-    /**
-     * @return array
-     */
     public function toArray(): array
     {
         if (! $this->isAuthorized()) {
