@@ -21,14 +21,18 @@ use Yajra\DataTables\Utilities\Helper;
  */
 class Editor extends Fluent
 {
-    use HasEvents;
     use HasAuthorizations;
+    use HasEvents;
 
-    const DISPLAY_LIGHTBOX = 'lightbox';
-    const DISPLAY_ENVELOPE = 'envelope';
-    const DISPLAY_BOOTSTRAP = 'bootstrap';
-    const DISPLAY_FOUNDATION = 'foundation';
-    const DISPLAY_JQUERYUI = 'jqueryui';
+    final public const DISPLAY_LIGHTBOX = 'lightbox';
+
+    final public const DISPLAY_ENVELOPE = 'envelope';
+
+    final public const DISPLAY_BOOTSTRAP = 'bootstrap';
+
+    final public const DISPLAY_FOUNDATION = 'foundation';
+
+    final public const DISPLAY_JQUERYUI = 'jqueryui';
 
     public array $events = [];
 
@@ -39,6 +43,7 @@ class Editor extends Fluent
      */
     public function __construct($instance = 'editor')
     {
+        $attributes = [];
         $attributes['instance'] = $instance;
 
         parent::__construct($attributes);
@@ -46,9 +51,6 @@ class Editor extends Fluent
 
     /**
      * Make new Editor instance.
-     *
-     * @param  array|string  $instance
-     * @return static
      */
     public static function make(array|string $instance = 'editor'): static
     {
@@ -62,7 +64,6 @@ class Editor extends Fluent
     /**
      * Append raw scripts.
      *
-     * @param  string  $scripts
      * @return $this
      */
     public function scripts(string $scripts): static
@@ -75,7 +76,6 @@ class Editor extends Fluent
     /**
      * Set Editor's variable name / instance.
      *
-     * @param  string  $instance
      * @return $this
      */
     public function instance(string $instance): static
@@ -88,8 +88,8 @@ class Editor extends Fluent
     /**
      * Set Editor's ajax parameter.
      *
-     * @param  array|string  $ajax
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/ajax
      */
     public function ajax(array|string $ajax): static
@@ -102,8 +102,8 @@ class Editor extends Fluent
     /**
      * Set Editor's table source.
      *
-     * @param  string  $table
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/table
      */
     public function table(string $table): static
@@ -116,8 +116,8 @@ class Editor extends Fluent
     /**
      * Set Editor's idSrc option.
      *
-     * @param  string  $idSrc
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/idSrc
      */
     public function idSrc(string $idSrc = 'DT_RowId'): static
@@ -130,8 +130,8 @@ class Editor extends Fluent
     /**
      * Set Editor's display option.
      *
-     * @param  string  $display
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/display
      */
     public function display(string $display): static
@@ -144,8 +144,8 @@ class Editor extends Fluent
     /**
      * Set Editor's fields.
      *
-     * @param  array  $fields
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields
      */
     public function fields(array $fields): static
@@ -158,8 +158,8 @@ class Editor extends Fluent
     /**
      * Set Editor's bubble formOptions.
      *
-     * @param  array|FormOptions  $formOptions
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/formOptions.bubble
      */
     public function formOptionsBubble(array|FormOptions $formOptions): static
@@ -170,8 +170,8 @@ class Editor extends Fluent
     /**
      * Set Editor's formOptions.
      *
-     * @param  array  $formOptions
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/formOptions
      * @see https://editor.datatables.net/reference/type/form-options
      */
@@ -185,8 +185,8 @@ class Editor extends Fluent
     /**
      * Set Editor's inline formOptions.
      *
-     * @param  array|FormOptions  $formOptions
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/formOptions.inline
      */
     public function formOptionsInline(array|FormOptions $formOptions): static
@@ -197,8 +197,8 @@ class Editor extends Fluent
     /**
      * Set Editor's main formOptions.
      *
-     * @param  array|FormOptions  $formOptions
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/formOptions.main
      */
     public function formOptionsMain(array|FormOptions $formOptions): static
@@ -209,8 +209,8 @@ class Editor extends Fluent
     /**
      * Set Editor's language.
      *
-     * @param  array  $language
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/i18n
      */
     public function language(array $language): static
@@ -223,8 +223,8 @@ class Editor extends Fluent
     /**
      * Set Editor's template.
      *
-     * @param  string  $template
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/template
      */
     public function template(string $template): static
@@ -236,8 +236,6 @@ class Editor extends Fluent
 
     /**
      * Convert the fluent instance to an array.
-     *
-     * @return array
      */
     public function toArray(): array
     {
@@ -262,7 +260,6 @@ class Editor extends Fluent
      * Convert the fluent instance to JSON.
      *
      * @param  int  $options
-     * @return string
      */
     public function toJson($options = 0): string
     {
@@ -276,7 +273,6 @@ class Editor extends Fluent
     /**
      * Hide fields on create action.
      *
-     * @param  array  $fields
      * @return $this
      */
     public function hiddenOnCreate(array $fields): static
@@ -287,8 +283,6 @@ class Editor extends Fluent
     /**
      * Hide fields on specific action.
      *
-     * @param  string  $action
-     * @param  array  $fields
      * @return $this
      */
     public function hiddenOn(string $action, array $fields): static
@@ -314,7 +308,6 @@ class Editor extends Fluent
     /**
      * Hide fields on edit action.
      *
-     * @param  array  $fields
      * @return $this
      */
     public function hiddenOnEdit(array $fields): static
