@@ -9,12 +9,8 @@ class BelongsTo extends Select
 {
     /**
      * @param  class-string<\Illuminate\Database\Eloquent\Model>|Builder  $class
-     * @param  string  $text
-     * @param  string  $id
-     * @param  string|null  $foreign
-     * @return static
      */
-    public static function model(Builder|string $class, string $text, string $id = 'id', string $foreign = null): static
+    public static function model(Builder|string $class, string $text, string $id = 'id', ?string $foreign = null): static
     {
         if ($class instanceof Builder) {
             $table = $class->getModel()->getTable();
@@ -23,9 +19,9 @@ class BelongsTo extends Select
         }
 
         $table = Str::singular($table);
-        $foreign = $foreign ?? $table.'_id';
+        $foreign ??= $table.'_id';
 
         return self::make($foreign, Str::title($table))
-                   ->modelOptions($class, $text, $id);
+            ->modelOptions($class, $text, $id);
     }
 }
