@@ -8,10 +8,6 @@ class DateTime extends Field
 
     /**
      * Make a new instance of a field.
-     *
-     * @param  array|string  $name
-     * @param  string  $label
-     * @return static
      */
     public static function make(array|string $name, string $label = ''): static
     {
@@ -29,9 +25,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  \DateTime  $dateTime
-     * @param  string  $format
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/options-min-max.html
      */
     public function minDate(\DateTime $dateTime, string $format = 'Y-m-d'): static
@@ -40,9 +35,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  \DateTime  $dateTime
-     * @param  string  $format
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/options-min-max.html
      */
     public function maxDate(\DateTime $dateTime, string $format = 'Y-m-d'): static
@@ -51,8 +45,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  bool  $state
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/options-week-numbers.html
      */
     public function showWeekNumber(bool $state = true): static
@@ -61,8 +55,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  array  $days
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/options-disable-days.html
      */
     public function disableDays(array $days): static
@@ -71,8 +65,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  int  $minutes
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/time-increment.html
      */
     public function minutesIncrement(int $minutes): static
@@ -81,8 +75,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  int  $seconds
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/time-increment.html
      */
     public function secondsIncrement(int $seconds): static
@@ -91,8 +85,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  array  $hours
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/datetime.html
      */
     public function hoursAvailable(array $hours): static
@@ -101,8 +95,8 @@ class DateTime extends Field
     }
 
     /**
-     * @param  array  $minutes
      * @return $this
+     *
      * @see https://editor.datatables.net/examples/dates/datetime.html
      */
     public function minutesAvailable(array $minutes): static
@@ -115,8 +109,8 @@ class DateTime extends Field
      * value and also for sending to the server on form submission.
      * The formatting options are defined by Moment.js.
      *
-     * @param  string  $format
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/field/datetime#Options
      * @see https://momentjs.com/docs/#/displaying/format/
      */
@@ -134,8 +128,8 @@ class DateTime extends Field
      * potentially making typing mistakes, although note that it does also disallow
      * pasting of data.
      *
-     * @param  bool  $state
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/field/datetime#Options
      */
     public function keyInput(bool $state = true): static
@@ -152,8 +146,8 @@ class DateTime extends Field
      * throw an error stating that Moment.js must be included for custom
      * formatting to be used.
      *
-     * @param  string  $format
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/field/datetime#Options
      * @see https://momentjs.com/docs/#/displaying/format/
      */
