@@ -21,8 +21,6 @@ class Field extends Fluent
 
     /**
      * Field type.
-     *
-     * @var string
      */
     protected string $type = 'text';
 
@@ -33,17 +31,13 @@ class Field extends Fluent
      */
     public function __construct($attributes = [])
     {
-        $attributes['type'] = $attributes['type'] ?? $this->type;
+        $attributes['type'] ??= $this->type;
 
         parent::__construct($attributes);
     }
 
     /**
      * Make a new instance of a field.
-     *
-     * @param  array|string  $name
-     * @param  string  $label
-     * @return static
      */
     public static function make(array|string $name, string $label = ''): static
     {
@@ -60,8 +54,8 @@ class Field extends Fluent
     }
 
     /**
-     * @param  string  $label
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.label
      */
     public function label(string $label): static
@@ -72,8 +66,8 @@ class Field extends Fluent
     }
 
     /**
-     * @param  string  $name
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.name
      */
     public function name(string $name): static
@@ -84,8 +78,8 @@ class Field extends Fluent
     }
 
     /**
-     * @param  string  $data
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.data
      */
     public function data(string $data): static
@@ -96,8 +90,8 @@ class Field extends Fluent
     }
 
     /**
-     * @param  string  $type
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.type
      */
     public function type(string $type): static
@@ -112,8 +106,6 @@ class Field extends Fluent
      * Get options from a model.
      *
      * @param  \Illuminate\Database\Eloquent\Builder|class-string<\Illuminate\Database\Eloquent\Model>  $model
-     * @param  string  $value
-     * @param  string  $key
      * @return $this
      */
     public function modelOptions(Builder|string $model, string $value, string $key = 'id'): static
@@ -126,7 +118,6 @@ class Field extends Fluent
     /**
      * Get options from a Enum::cases().
      *
-     * @param array $cases
      * @return $this
      */
     public function enumOptions(array $cases): static
@@ -145,7 +136,6 @@ class Field extends Fluent
     /**
      * Set select options.
      *
-     * @param  array|Arrayable  $options
      * @return $this
      */
     public function options(array|Arrayable $options): static
@@ -162,19 +152,14 @@ class Field extends Fluent
     /**
      * Get options from a table.
      *
-     * @param  QueryBuilder|\Closure|string  $table
-     * @param  string  $value
-     * @param  string  $key
-     * @param  \Closure|null  $whereCallback
-     * @param  string|null  $connection
      * @return $this
      */
     public function tableOptions(
         QueryBuilder|Closure|string $table,
         string $value,
         string $key = 'id',
-        Closure $whereCallback = null,
-        string $connection = null
+        ?Closure $whereCallback = null,
+        ?string $connection = null
     ): static {
         return $this->options(
             Options::table($table, $value, $key, $whereCallback, $connection)
@@ -184,7 +169,6 @@ class Field extends Fluent
     /**
      * Set checkbox separator.
      *
-     * @param  string  $separator
      * @return $this
      */
     public function separator(string $separator = ','): static
@@ -197,8 +181,8 @@ class Field extends Fluent
     /**
      * Set dateTime format.
      *
-     * @param  string  $format
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/field/datetime
      */
     public function format(string $format): static
@@ -211,8 +195,8 @@ class Field extends Fluent
     /**
      * Set field default value.
      *
-     * @param  float|bool|int|string|array  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.def
      */
     public function default(float|bool|int|string|array $value): static
@@ -225,8 +209,8 @@ class Field extends Fluent
     /**
      * Set field message value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.message
      */
     public function message(string $value): static
@@ -239,8 +223,8 @@ class Field extends Fluent
     /**
      * Set field fieldInfo value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.fieldInfo
      */
     public function fieldInfo(string $value): static
@@ -253,8 +237,8 @@ class Field extends Fluent
     /**
      * Set field labelInfo value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.labelInfo
      */
     public function labelInfo(string $value): static
@@ -267,8 +251,8 @@ class Field extends Fluent
     /**
      * Set field entityDecode value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.entityDecode
      */
     public function entityDecode(bool $value): static
@@ -281,8 +265,8 @@ class Field extends Fluent
     /**
      * Set field multiEditable value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.multiEditable
      */
     public function multiEditable(bool $value): static
@@ -295,8 +279,8 @@ class Field extends Fluent
     /**
      * Set field id value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.id
      */
     public function id(string $value): static
@@ -309,8 +293,8 @@ class Field extends Fluent
     /**
      * Set field submit value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.submit
      */
     public function submit(bool $value): static
@@ -323,8 +307,8 @@ class Field extends Fluent
     /**
      * Set field compare value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.compare
      */
     public function compare(bool $value): static
@@ -337,8 +321,8 @@ class Field extends Fluent
     /**
      * Set field opts value.
      *
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/forums/discussion/comment/156581/#Comment_156581
      */
     public function opts(array $value): static
@@ -355,9 +339,8 @@ class Field extends Fluent
     /**
      * Set field element html attributes.
      *
-     * @param  string  $attribute
-     * @param  int|bool|string  $value
      * @return $this
+     *
      * @see https://datatables.net/forums/discussion/comment/156581/#Comment_156581
      */
     public function attr(string $attribute, int|bool|string $value): static
@@ -374,9 +357,6 @@ class Field extends Fluent
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getType(): string
     {
         return $this->type;
@@ -385,8 +365,8 @@ class Field extends Fluent
     /**
      * Replace null values with the field's default on edit.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/option/fields.nullDefault
      */
     public function nullDefault(bool $value = true): static
