@@ -10,7 +10,6 @@ class Select2 extends Select
     protected string $type = 'select2';
 
     /**
-     * @param  bool  $state
      * @return $this
      */
     public function allowClear(bool $state = true): static
@@ -19,7 +18,6 @@ class Select2 extends Select
     }
 
     /**
-     * @param  string  $value
      * @return $this
      */
     public function placeholder(string $value): static
@@ -28,8 +26,6 @@ class Select2 extends Select
     }
 
     /**
-     * @param  string  $text
-     * @param  string  $id
      * @return $this
      */
     public function optsPlaceholder(string $text = '', string $id = ''): static
@@ -45,7 +41,6 @@ class Select2 extends Select
     /**
      * Set select2 ajax option.
      *
-     * @param  array|string  $value
      * @return $this
      */
     public function ajax(array|string $value): static
@@ -62,7 +57,6 @@ class Select2 extends Select
     /**
      * Set select2 ajax url option.
      *
-     * @param  string  $value
      * @return $this
      */
     public function ajaxUrl(string $value): static
@@ -73,7 +67,6 @@ class Select2 extends Select
     /**
      * Set select2 ajaxDelay option.
      *
-     * @param  int  $value
      * @return $this
      */
     public function ajaxDelay(int $value = 250): static
@@ -84,7 +77,6 @@ class Select2 extends Select
     /**
      * Set select2 ajax data option.
      *
-     * @param  array|string  $data
      * @return $this
      */
     public function ajaxData(array|string $data): static
@@ -92,7 +84,7 @@ class Select2 extends Select
         if (is_array($data)) {
             $script = 'function(params) {';
             foreach ($data as $key => $value) {
-                $value = json_encode($value);
+                $value = json_encode($value, JSON_THROW_ON_ERROR);
                 $script .= " params.$key = $value; ";
             }
             $script .= 'return params; }';
@@ -106,9 +98,6 @@ class Select2 extends Select
     /**
      * Set select2 ajax processResults option to process a paginated results.
      *
-     * @param  string  $display
-     * @param  string  $id
-     * @param  string  $wrap
      * @return $this
      */
     public function processPaginatedResults(string $display = 'text', string $id = 'id', string $wrap = 'results'): static
@@ -125,7 +114,6 @@ class Select2 extends Select
     /**
      * Set select2 ajax processResults option.
      *
-     * @param  string  $value
      * @return $this
      */
     public function processResults(string $value): static
