@@ -9,21 +9,17 @@ use Illuminate\Support\Fluent;
  */
 class FormOptions extends Fluent
 {
-    /**
-     * @param  array  $attributes
-     * @return static
-     */
     public static function make(array $attributes = []): static
     {
         return new static($attributes);
     }
 
     /**
-     * @param  int|string|null  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#focus
      */
-    public function focus(int|string $value = null): static
+    public function focus(int|string|null $value = null): static
     {
         $this->attributes['focus'] = $value;
 
@@ -31,8 +27,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#nest
      */
     public function nest(bool $value): static
@@ -43,8 +39,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#onBackground
      */
     public function onBackground(string $value = 'blur'): static
@@ -55,8 +51,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#onBlur
      */
     public function onBlur(string $value = 'close'): static
@@ -67,8 +63,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#onComplete
      */
     public function onComplete(string $value = 'close'): static
@@ -79,8 +75,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#onEsc
      */
     public function onEsc(string $value = 'close'): static
@@ -91,8 +87,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#onFieldError
      */
     public function onFieldError(string $value = 'focus'): static
@@ -103,8 +99,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#onReturn
      */
     public function onReturn(string $value = 'submit'): static
@@ -115,8 +111,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#submit
      */
     public function submit(string $value = 'changed'): static
@@ -127,11 +123,11 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#scope
      */
-    public function scope(string $value = 'row'): static
+    public function formScope(string $value = 'row'): static
     {
         $this->attributes['scope'] = $value;
 
@@ -139,8 +135,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  array|string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#buttons
      */
     public function buttons(array|string $value): static
@@ -149,9 +145,10 @@ class FormOptions extends Fluent
 
         return $this;
     }
+
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#drawType
      */
     public function drawType(string $value = ''): static
@@ -162,8 +159,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  bool|string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#message
      */
     public function message(bool|string $value = ''): static
@@ -174,8 +171,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  int|string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#submitTrigger
      */
     public function submitTrigger(int|string $value): static
@@ -186,8 +183,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#submitHtml
      */
     public function submitHtml(string $value): static
@@ -198,8 +195,8 @@ class FormOptions extends Fluent
     }
 
     /**
-     * @param  bool|string  $value
      * @return $this
+     *
      * @see https://editor.datatables.net/reference/type/form-options#title
      */
     public function title(bool|string $value): static
