@@ -8,17 +8,11 @@ trait HasAuthorizations
 {
     /**
      * Flag to check if user is authorized to use the button.
-     *
-     * @var bool
      */
     protected bool $authorized = true;
 
     /**
      * Make a button if condition is true.
-     *
-     * @param  callable|bool  $condition
-     * @param  array|string  $options
-     * @return static
      */
     public static function makeIf(callable|bool $condition, array|string $options = []): static
     {
@@ -35,9 +29,6 @@ trait HasAuthorizations
 
     /**
      * Set authorization status of the button.
-     *
-     * @param  callable|bool  $bool
-     * @return static
      */
     public function authorized(callable|bool $bool): static
     {
@@ -48,13 +39,8 @@ trait HasAuthorizations
 
     /**
      * Make a button if the user is authorized.
-     *
-     * @param  string  $permission
-     * @param  array|string  $options
-     * @param  Authorizable|null  $user
-     * @return static
      */
-    public static function makeIfCan(string $permission, array|string $options = [], Authorizable $user = null): static
+    public static function makeIfCan(string $permission, array|string $options = [], ?Authorizable $user = null): static
     {
         if (is_null($user)) {
             $user = auth()->user();
@@ -69,16 +55,11 @@ trait HasAuthorizations
 
     /**
      * Make a button if the user is not authorized.
-     *
-     * @param  string  $permission
-     * @param  array|string  $options
-     * @param  Authorizable|null  $user
-     * @return static
      */
     public static function makeIfCannot(
         string $permission,
         array|string $options = [],
-        Authorizable $user = null
+        ?Authorizable $user = null
     ): static {
         if (is_null($user)) {
             $user = auth()->user();
@@ -93,8 +74,6 @@ trait HasAuthorizations
 
     /**
      * Convert the Fluent instance to an array.
-     *
-     * @return array
      */
     public function toArray(): array
     {
@@ -107,8 +86,6 @@ trait HasAuthorizations
 
     /**
      * Check if instance is authorized
-     *
-     * @return bool
      */
     public function isAuthorized(): bool
     {
