@@ -8,8 +8,6 @@ trait HasEditor
 {
     /**
      * Collection of Editors.
-     *
-     * @var array
      */
     protected array $editors = [];
 
@@ -18,6 +16,7 @@ trait HasEditor
      *
      * @param  array|mixed  ...$editors
      * @return $this
+     *
      * @see https://editor.datatables.net/
      */
     public function editors(...$editors): static
@@ -42,8 +41,8 @@ trait HasEditor
     /**
      * Integrate with DataTables Editor.
      *
-     * @param  Editor  $editor
      * @return $this
+     *
      * @see https://editor.datatables.net/
      */
     public function editor(Editor $editor): static
@@ -66,9 +65,6 @@ trait HasEditor
         return $this;
     }
 
-    /**
-     * @return array
-     */
     public function getEditors(): array
     {
         return $this->editors;
