@@ -10,8 +10,6 @@ trait HasTable
 
     /**
      * Get table computed table attributes.
-     *
-     * @return array
      */
     public function getTableAttributes(): array
     {
@@ -21,7 +19,6 @@ trait HasTable
     /**
      * Sets HTML table "id" attribute.
      *
-     * @param  string  $id
      * @return $this
      */
     public function setTableId(string $id): static
@@ -32,11 +29,9 @@ trait HasTable
     /**
      * Sets HTML table attribute(s).
      *
-     * @param  array|string  $attribute
-     * @param  string|null  $value
      * @return $this
      */
-    public function setTableAttribute(array|string $attribute, string $value = null): static
+    public function setTableAttribute(array|string $attribute, ?string $value = null): static
     {
         if (is_array($attribute)) {
             return $this->setTableAttributes($attribute);
@@ -50,7 +45,6 @@ trait HasTable
     /**
      * Sets multiple HTML table attributes at once.
      *
-     * @param  array  $attributes
      * @return $this
      */
     public function setTableAttributes(array $attributes): static
@@ -64,8 +58,6 @@ trait HasTable
 
     /**
      * Get HTML table "id" attribute.
-     *
-     * @return string
      */
     public function getTableId(): string
     {
@@ -74,9 +66,6 @@ trait HasTable
 
     /**
      * Retrieves HTML table attribute value.
-     *
-     * @param  string  $attribute
-     * @return string
      */
     public function getTableAttribute(string $attribute): string
     {
@@ -86,7 +75,6 @@ trait HasTable
     /**
      * Add class names to the "class" attribute of HTML table.
      *
-     * @param  array|string  $class
      * @return $this
      */
     public function addTableClass(array|string $class): static
@@ -103,7 +91,6 @@ trait HasTable
     /**
      * Set table > thead class names.
      *
-     * @param  string  $class
      * @return $this
      */
     public function setTableHeadClass(string $class): static
@@ -116,7 +103,6 @@ trait HasTable
     /**
      * Remove class names from the "class" attribute of HTML table.
      *
-     * @param  array|string  $class
      * @return $this
      */
     public function removeTableClass(array|string $class): static
@@ -125,7 +111,7 @@ trait HasTable
         $currentClass = $this->getTableAttribute('class');
 
         $classes = array_diff(
-            (array) preg_split('#\s+#', $currentClass, -1, PREG_SPLIT_NO_EMPTY),
+            (array) preg_split('#\s+#', (string) $currentClass, -1, PREG_SPLIT_NO_EMPTY),
             (array) preg_split('#\s+#', $class, -1, PREG_SPLIT_NO_EMPTY)
         );
         $class = implode(' ', array_unique($classes));
@@ -135,8 +121,6 @@ trait HasTable
 
     /**
      * Compile table headers and to support responsive extension.
-     *
-     * @return array
      */
     protected function compileTableHeaders(): array
     {
@@ -163,8 +147,6 @@ trait HasTable
 
     /**
      * Compile table search headers.
-     *
-     * @return array
      */
     protected function compileTableSearchHeaders(): array
     {
@@ -179,8 +161,6 @@ trait HasTable
 
     /**
      * Compile table footer contents.
-     *
-     * @return array
      */
     protected function compileTableFooter(): array
     {
