@@ -2,8 +2,6 @@
 
 namespace Yajra\DataTables\Html\Options;
 
-use Yajra\DataTables\Html\Options\Languages;
-
 /**
  * DataTables - Internationalisation option builder.
  *
@@ -19,8 +17,8 @@ trait HasInternationalisation
     /**
      * Set language decimal option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.decimal
      */
     public function languageDecimal(string $value): static
@@ -31,8 +29,8 @@ trait HasInternationalisation
     /**
      * Set language option value.
      *
-     * @param  array|string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language
      */
     public function language(array|string $value): static
@@ -49,8 +47,8 @@ trait HasInternationalisation
     /**
      * Set language emptyTable option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.emptyTable
      */
     public function languageEmptyTable(string $value): static
@@ -61,8 +59,8 @@ trait HasInternationalisation
     /**
      * Set language info option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.info
      */
     public function languageInfo(string $value): static
@@ -73,8 +71,8 @@ trait HasInternationalisation
     /**
      * Set language infoEmpty option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.infoEmpty
      */
     public function languageInfoEmpty(string $value): static
@@ -85,8 +83,8 @@ trait HasInternationalisation
     /**
      * Set language infoFiltered option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.infoFiltered
      */
     public function languageInfoFiltered(string $value): static
@@ -97,8 +95,8 @@ trait HasInternationalisation
     /**
      * Set language infoPostFix option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.infoPostFix
      */
     public function languageInfoPostFix(string $value): static
@@ -109,8 +107,8 @@ trait HasInternationalisation
     /**
      * Set language lengthMenu option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.lengthMenu
      */
     public function languageLengthMenu(string $value): static
@@ -121,8 +119,8 @@ trait HasInternationalisation
     /**
      * Set language loadingRecords option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.loadingRecords
      */
     public function languageLoadingRecords(string $value): static
@@ -133,8 +131,8 @@ trait HasInternationalisation
     /**
      * Set language processing option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.processing
      */
     public function languageProcessing(string $value): static
@@ -145,8 +143,8 @@ trait HasInternationalisation
     /**
      * Set language search option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.search
      */
     public function languageSearch(string $value): static
@@ -157,8 +155,8 @@ trait HasInternationalisation
     /**
      * Set language searchPlaceholder option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.searchPlaceholder
      */
     public function languageSearchPlaceholder(string $value): static
@@ -169,8 +167,8 @@ trait HasInternationalisation
     /**
      * Set language thousands option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.thousands
      */
     public function languageThousands(string $value): static
@@ -181,8 +179,8 @@ trait HasInternationalisation
     /**
      * Set language url option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.url
      */
     public function languageUrl(string $value): static
@@ -193,8 +191,8 @@ trait HasInternationalisation
     /**
      * Set language zeroRecords option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.zeroRecords
      */
     public function languageZeroRecords(string $value): static
@@ -202,11 +200,7 @@ trait HasInternationalisation
         return $this->language(['zeroRecords' => $value]);
     }
 
-    /**
-     * @param  string|null  $key
-     * @return mixed
-     */
-    public function getLanguage(string $key = null): mixed
+    public function getLanguage(?string $key = null): mixed
     {
         if (is_null($key)) {
             return $this->attributes['language'] ?? [];
