@@ -7,8 +7,8 @@ trait Aria
     /**
      * Set language aria option value.
      *
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria
      */
     public function languageAria(array $value): static
@@ -19,8 +19,8 @@ trait Aria
     /**
      * Set language aria paginate option value.
      *
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.paginate
      */
     public function languageAriaPaginate(array $value): static
@@ -31,8 +31,8 @@ trait Aria
     /**
      * Set language aria paginate first option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.paginate.first
      */
     public function languageAriaPaginateFirst(string $value): static
@@ -43,8 +43,8 @@ trait Aria
     /**
      * Set language aria paginate last option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.paginate.last
      */
     public function languageAriaPaginateLast(string $value): static
@@ -55,8 +55,8 @@ trait Aria
     /**
      * Set language aria paginate next option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.paginate.next
      */
     public function languageAriaPaginateNext(string $value): static
@@ -67,8 +67,8 @@ trait Aria
     /**
      * Set language aria paginate previous option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.paginate.previous
      */
     public function languageAriaPaginatePrevious(string $value): static
@@ -79,8 +79,8 @@ trait Aria
     /**
      * Set language aria sortAscending option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.sortAscending
      */
     public function languageAriaSortAscending(string $value): static
@@ -91,8 +91,8 @@ trait Aria
     /**
      * Set language aria sortDescending option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/language.aria.sortDescending
      */
     public function languageAriaSortDescending(string $value): static
