@@ -14,8 +14,8 @@ trait FixedColumns
     /**
      * Set fixedColumns heightMatch option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedColumns.heightMatch
      */
     public function fixedColumnsHeightMatch(string $value = 'semiauto'): static
@@ -26,8 +26,8 @@ trait FixedColumns
     /**
      * Set fixedColumns option value.
      *
-     * @param  array|bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedColumns
      */
     public function fixedColumns(array|bool $value = true): static
@@ -38,8 +38,8 @@ trait FixedColumns
     /**
      * Set fixedColumns leftColumns option value.
      *
-     * @param  int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedColumns.leftColumns
      */
     public function fixedColumnsLeftColumns(int $value = 1): static
@@ -50,8 +50,8 @@ trait FixedColumns
     /**
      * Set fixedColumns rightColumns option value.
      *
-     * @param  int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedColumns.rightColumns
      */
     public function fixedColumnsRightColumns(int $value = 0): static
@@ -59,11 +59,7 @@ trait FixedColumns
         return $this->fixedColumns(['rightColumns' => $value]);
     }
 
-    /**
-     * @param  string|null  $key
-     * @return mixed
-     */
-    public function getFixedColumns(string $key = null): mixed
+    public function getFixedColumns(?string $key = null): mixed
     {
         if (is_null($key)) {
             return $this->attributes['fixedColumns'] ?? true;
