@@ -14,8 +14,8 @@ trait FixedHeader
     /**
      * Set fixedHeader footer option value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedHeader.footer
      */
     public function fixedHeaderFooter(bool $value = true): static
@@ -26,20 +26,20 @@ trait FixedHeader
     /**
      * Set fixedHeader option value.
      *
-     * @param  array|bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedHeader
      */
     public function fixedHeader(array|bool $value = true): static
     {
-        return  $this->setPluginAttribute('fixedHeader', $value);
+        return $this->setPluginAttribute('fixedHeader', $value);
     }
 
     /**
      * Set fixedHeader footerOffset option value.
      *
-     * @param  int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedHeader.footerOffset
      */
     public function fixedHeaderFooterOffset(int $value = 0): static
@@ -50,8 +50,8 @@ trait FixedHeader
     /**
      * Set fixedHeader header option value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedHeader.header
      */
     public function fixedHeaderHeader(bool $value = true): static
@@ -62,8 +62,8 @@ trait FixedHeader
     /**
      * Set fixedHeader headerOffset option value.
      *
-     * @param  int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/fixedHeader.headerOffset
      */
     public function fixedHeaderHeaderOffset(int $value = 0): static
@@ -71,11 +71,7 @@ trait FixedHeader
         return $this->fixedHeader(['headerOffset' => $value]);
     }
 
-    /**
-     * @param  string|null  $key
-     * @return mixed
-     */
-    public function getFixedHeader(string $key = null): mixed
+    public function getFixedHeader(?string $key = null): mixed
     {
         if (is_null($key)) {
             return $this->attributes['fixedHeader'] ?? true;
