@@ -14,8 +14,8 @@ trait RowReorder
     /**
      * Set rowReorder dataSrc option value.
      *
-     * @param  array|int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.dataSrc
      */
     public function rowReorderDataSrc(array|int $value = 0): static
@@ -26,8 +26,8 @@ trait RowReorder
     /**
      * Set rowReorder option value.
      *
-     * @param  array|bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder
      */
     public function rowReorder(array|bool $value = true): static
@@ -38,11 +38,11 @@ trait RowReorder
     /**
      * Set rowReorder editor option value.
      *
-     * @param  string|null  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.editor
      */
-    public function rowReorderEditor(string $value = null): static
+    public function rowReorderEditor(?string $value = null): static
     {
         return $this->rowReorder(['editor' => $value]);
     }
@@ -50,8 +50,8 @@ trait RowReorder
     /**
      * Set rowReorder enable option value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.enable
      */
     public function rowReorderEnable(bool $value = true): static
@@ -62,8 +62,8 @@ trait RowReorder
     /**
      * Set rowReorder formOptions option value.
      *
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.formOptions
      */
     public function rowReorderFormOptions(array $value): static
@@ -74,8 +74,8 @@ trait RowReorder
     /**
      * Set rowReorder selector option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.selector
      */
     public function rowReorderSelector(string $value = 'td:first-child'): static
@@ -86,8 +86,8 @@ trait RowReorder
     /**
      * Set rowReorder snapX option value.
      *
-     * @param  bool|int  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.snapX
      */
     public function rowReorderSnapX(bool|int $value = true): static
@@ -98,8 +98,8 @@ trait RowReorder
     /**
      * Set rowReorder update option value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/rowReorder.update
      */
     public function rowReorderUpdate(bool $value = true): static
@@ -107,11 +107,7 @@ trait RowReorder
         return $this->rowReorder(['update' => $value]);
     }
 
-    /**
-     * @param  string|null  $key
-     * @return mixed
-     */
-    public function getRowReorder(string $key = null): mixed
+    public function getRowReorder(?string $key = null): mixed
     {
         if (is_null($key)) {
             return $this->attributes['rowReorder'] ?? true;
