@@ -15,8 +15,8 @@ trait Select
     /**
      * Set select blurable option value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.blurable
      */
     public function selectBlurable(bool $value = true): static
@@ -27,8 +27,8 @@ trait Select
     /**
      * Set select className option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.className
      */
     public function selectClassName(string $value = 'selected'): static
@@ -39,7 +39,6 @@ trait Select
     /**
      * Append a class name to className option value.
      *
-     * @param  string  $class
      * @return $this
      */
     public function selectAddClassName(string $class): static
@@ -56,8 +55,8 @@ trait Select
     /**
      * Set select info option value.
      *
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.info
      */
     public function selectInfo(bool $value = true): static
@@ -68,8 +67,8 @@ trait Select
     /**
      * Set select items option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.items
      */
     public function selectItems(string $value = 'row'): static
@@ -81,6 +80,7 @@ trait Select
      * Set select items option value to row.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.items
      */
     public function selectItemsRow(): static
@@ -92,6 +92,7 @@ trait Select
      * Set select items option value to column.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.items
      */
     public function selectItemsColumn(): static
@@ -103,6 +104,7 @@ trait Select
      * Set select items option value to cell.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.items
      */
     public function selectItemsCell(): static
@@ -113,8 +115,8 @@ trait Select
     /**
      * Set select selector option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.selector
      */
     public function selectSelector(string $value = 'td'): static
@@ -125,8 +127,8 @@ trait Select
     /**
      * Set select style option value.
      *
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.style
      */
     public function selectStyle(string $value = 'os'): static
@@ -138,6 +140,7 @@ trait Select
      * Set select style option value to api.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.style
      */
     public function selectStyleApi(): static
@@ -149,6 +152,7 @@ trait Select
      * Set select style option value to single.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.style
      */
     public function selectStyleSingle(): static
@@ -159,8 +163,8 @@ trait Select
     /**
      * Set select option value.
      *
-     * @param  array|bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select
      */
     public function select(bool|array $value = true): static
@@ -172,6 +176,7 @@ trait Select
      * Set select style option value to multi.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.style
      */
     public function selectStyleMulti(): static
@@ -183,6 +188,7 @@ trait Select
      * Set select style option value to os.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.style
      */
     public function selectStyleOS(): static
@@ -194,6 +200,7 @@ trait Select
      * Set select style option value to multi+shift.
      *
      * @return $this
+     *
      * @see https://datatables.net/reference/option/select.style
      */
     public function selectStyleMultiShift(): static
@@ -201,11 +208,7 @@ trait Select
         return $this->select(['style' => Builder::SELECT_STYLE_MULTI_SHIFT]);
     }
 
-    /**
-     * @param  string|null  $key
-     * @return mixed
-     */
-    public function getSelect(string $key = null): mixed
+    public function getSelect(?string $key = null): mixed
     {
         if (is_null($key)) {
             return $this->attributes['select'] ?? true;
