@@ -15,18 +15,14 @@ class SearchPane extends Fluent
         parent::__construct(['show' => true] + $attributes);
     }
 
-    /**
-     * @param  array  $options
-     * @return static
-     */
     public static function make(array $options = []): static
     {
         return new static($options);
     }
 
     /**
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.cascadePanes
      */
     public function cascadePanes(bool $value = true): static
@@ -37,8 +33,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.clear
      */
     public function clear(bool $value = true): static
@@ -49,8 +45,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.columns
      */
     public function columns(array $value = []): static
@@ -61,8 +57,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.controls
      */
     public function controls(bool $value = true): static
@@ -73,8 +69,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.dtOpts
      * @see https://datatables.net/reference/option/columns.searchPanes.dtOpts
      */
@@ -86,8 +82,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.emptyMessage
      */
     public function emptyMessage(string $value): static
@@ -98,8 +94,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.filterChanged
      */
     public function filterChanged(string $value): static
@@ -110,8 +106,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.hideCount
      */
     public function hideCount(bool $value = true): static
@@ -122,8 +118,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.layout
      */
     public function layout(string $value): static
@@ -134,8 +130,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.order
      */
     public function order(array $value): static
@@ -146,8 +142,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  boolean  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.orderable
      */
     public function orderable(bool $value = true): static
@@ -158,8 +154,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  array  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.panes
      */
     public function panes(array $value): static
@@ -178,8 +174,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  float  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.threshold
      */
     public function threshold(float $value): static
@@ -190,8 +186,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  boolean  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.viewTotal
      */
     public function viewTotal(bool $value = true): static
@@ -202,8 +198,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  boolean  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.viewTotal
      */
     public function hideTotal(bool $value = true): static
@@ -217,9 +213,6 @@ class SearchPane extends Fluent
      * Get options from a model.
      *
      * @param  class-string<\Illuminate\Database\Eloquent\Model>|EloquentBuilder  $model
-     * @param  string  $value
-     * @param  string  $key
-     * @return $this
      */
     public function modelOptions(EloquentBuilder|string $model, string $value, string $key = 'id'): SearchPane
     {
@@ -227,8 +220,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  array|Arrayable  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.searchPanes.options
      */
     public function options(array|Arrayable $value): static
@@ -245,26 +238,21 @@ class SearchPane extends Fluent
     /**
      * Get options from a table.
      *
-     * @param  string  $table
-     * @param  string  $value
-     * @param  string  $key
-     * @param  \Closure|null  $callback
-     * @param  string|null  $connection
      * @return $this
      */
     public function tableOptions(
         string $table,
         string $value,
         string $key = 'id',
-        Closure $callback = null,
-        string $connection = null
+        ?Closure $callback = null,
+        ?string $connection = null
     ): static {
         return $this->options(Options::table($table, $value, $key, $callback, $connection));
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.searchPanes.className
      */
     public function className(string $value): static
@@ -275,8 +263,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.panes.header
      */
     public function header(string $value): static
@@ -287,8 +275,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  bool  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.searchPanes.show
      */
     public function show(bool $value = true): static
@@ -299,8 +287,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.searchPanes.name
      */
     public function name(string $value): static
@@ -311,8 +299,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  array|string  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/columns.searchPanes.orthogonal
      */
     public function orthogonal(array|string $value): static
@@ -323,8 +311,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  boolean  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.collapse
      */
     public function collapse(bool $value = true): static
@@ -335,8 +323,8 @@ class SearchPane extends Fluent
     }
 
     /**
-     * @param  boolean  $value
      * @return $this
+     *
      * @see https://datatables.net/reference/option/searchPanes.initCollapsed
      */
     public function initCollapsed(bool $value = false): static
