@@ -12,7 +12,7 @@ return [
      */
     'table' => [
         'class' => 'table',
-        'id'    => 'dataTableBuilder',
+        'id' => 'dataTableBuilder',
     ],
 
     /*
